third_party/blink/web_tests/external/wpt/css/css-cascade/scope-import-implicit-expected.txt

This is a testharness.js-based test.
[FAIL] Scope-imported rule does not apply outside of implicit scope
  assert_equals: expected "" but got "1"
Harness: the test ran to completion.
